feat: pluginify services
