+++ packages/vscode/src/plugins/index.ts
@@ -1,0 +1,6 @@
+import languageServer from './prisma-language-server'
+import { PrismaVSCodePlugin } from './types'
+
+const plugins: PrismaVSCodePlugin[] = [languageServer]
+
+export default plugins
